Move C++ code to Utils.hpp
